requirements.txt
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
requests-ntlm==1.3.0
exchangelib==5.4.2
icalendar==6.1.3
